refactor(run): extract claim-lease + prev-failure lookup helpers to hold the process-deps line budget Refs #2174
